Consolidates manual URL utils in usermanual_url.
Uses development manual URL in welcome screen.
